# legacy-api/openapi.yaml
openapi: 3.0.3
info:
  title: Legacy Fixture API
  version: 0.1.0
  description: Correctly documented legacy-style API fixture for API probing and SDK generation tests.
servers:
  - url: http://localhost:4011
tags:
  - name: system
  - name: orders
  - name: invoices
  - name: expenses
paths:
  /health:
    get:
      tags: [system]
      summary: Health check
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    example: ok

  /api/{version}/system/info:
    get:
      tags: [system]
      summary: Get system metadata for a supported version
      parameters:
        - $ref: '#/components/parameters/version'
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: System information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemInfoResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /api/{version}/orders/template:
    get:
      tags: [orders]
      summary: Get a short-lived order template token required before creating or updating an order
      parameters:
        - $ref: '#/components/parameters/version'
        - name: customer_id
          in: query
          required: true
          schema:
            type: string
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Order template generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderTemplateResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /api/{version}/orders/{orderId}:
    put:
      tags: [orders]
      summary: Create or update an order
      description: |
        For version v1.1, the request body must include commodity_id.
        For version v1.2, the request body must include commodity_code_id.
      parameters:
        - $ref: '#/components/parameters/version'
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/OrderUpsertRequestV11'
                - $ref: '#/components/schemas/OrderUpsertRequestV12'
      responses:
        '200':
          description: Updated existing order
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/OrderResponseV11'
                  - $ref: '#/components/schemas/OrderResponseV12'
        '201':
          description: Created new order
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/OrderResponseV11'
                  - $ref: '#/components/schemas/OrderResponseV12'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    get:
      tags: [orders]
      summary: Fetch an order by ID
      parameters:
        - $ref: '#/components/parameters/version'
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/OrderResponseV11'
                  - $ref: '#/components/schemas/OrderResponseV12'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /api/{version}/invoices:
    post:
      tags: [invoices]
      summary: Create an invoice
      parameters:
        - $ref: '#/components/parameters/version'
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceUpsertRequest'
      responses:
        '201':
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /api/{version}/invoices/{invoiceId}:
    get:
      tags: [invoices]
      summary: Fetch invoice by ID
      parameters:
        - $ref: '#/components/parameters/version'
        - name: invoiceId
          in: path
          required: true
          schema:
            type: string
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Invoice found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /api/{version}/expenses/summary:
    get:
      tags: [expenses]
      summary: Get customer invoice totals for a month
      parameters:
        - $ref: '#/components/parameters/version'
        - name: customer_id
          in: query
          required: true
          schema:
            type: string
        - name: month
          in: query
          required: true
          schema:
            type: string
            pattern: '^\\d{4}-\\d{2}$'
            example: '2026-02'
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Expense summary response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseSummaryResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key

  parameters:
    version:
      name: version
      in: path
      required: true
      schema:
        type: string
        enum: [v1.1, v1.2]

  schemas:
    SystemInfoResponse:
      type: object
      required: [system, version, auth, orderCreateFlow, notes]
      properties:
        system:
          type: string
          example: Legacy Freight Suite
        version:
          type: string
          enum: [v1.1, v1.2]
        auth:
          type: object
          required: [header]
          properties:
            header:
              type: string
              example: x-api-key
        orderCreateFlow:
          type: array
          items:
            type: string
        notes:
          type: string

    OrderTemplateResponse:
      type: object
      required: [template_token, customer_id, expires_at, default_currency, required_fields]
      properties:
        template_token:
          type: string
        customer_id:
          type: string
        expires_at:
          type: string
          format: date-time
        default_currency:
          type: string
          example: USD
        required_fields:
          type: array
          items:
            type: string

    LineItem:
      type: object
      required: [sku, quantity, unit_price]
      properties:
        sku:
          type: string
        quantity:
          type: number
          minimum: 0.00001
        unit_price:
          type: number
          minimum: 0

    OrderUpsertRequestBase:
      type: object
      required: [template_token, customer_id, pickup_date, line_items]
      properties:
        template_token:
          type: string
        customer_id:
          type: string
        pickup_date:
          type: string
          description: Date string expected by legacy system
          example: '2026-03-14'
        currency:
          type: string
          default: USD
        line_items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/LineItem'

    OrderUpsertRequestV11:
      allOf:
        - $ref: '#/components/schemas/OrderUpsertRequestBase'
        - type: object
          required: [commodity_id]
          properties:
            commodity_id:
              type: string

    OrderUpsertRequestV12:
      allOf:
        - $ref: '#/components/schemas/OrderUpsertRequestBase'
        - type: object
          required: [commodity_code_id]
          properties:
            commodity_code_id:
              type: string

    OrderResponseBase:
      type: object
      required: [order_id, customer_id, pickup_date, currency, line_items, created_at, updated_at]
      properties:
        order_id:
          type: string
        customer_id:
          type: string
        pickup_date:
          type: string
        currency:
          type: string
        line_items:
          type: array
          items:
            $ref: '#/components/schemas/LineItem'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    OrderResponseV11:
      allOf:
        - $ref: '#/components/schemas/OrderResponseBase'
        - type: object
          required: [commodity_id]
          properties:
            commodity_id:
              type: string

    OrderResponseV12:
      allOf:
        - $ref: '#/components/schemas/OrderResponseBase'
        - type: object
          required: [commodity_code_id]
          properties:
            commodity_code_id:
              type: string

    InvoiceUpsertRequest:
      type: object
      required: [invoice_id, customer_id, total, currency, issued_at]
      properties:
        invoice_id:
          type: string
        customer_id:
          type: string
        total:
          type: number
        currency:
          type: string
          example: USD
        issued_at:
          type: string
          format: date-time
        notes:
          type: string

    InvoiceResponse:
      type: object
      required: [invoice_id, customer_id, total, currency, issued_at]
      properties:
        invoice_id:
          type: string
        customer_id:
          type: string
        total:
          type: number
        currency:
          type: string
        issued_at:
          type: string
          format: date-time
        notes:
          type: string

    ExpenseSummaryResponse:
      type: object
      required: [customer_id, month, invoice_count, total_amount, currency]
      properties:
        customer_id:
          type: string
        month:
          type: string
          pattern: '^\\d{4}-\\d{2}$'
        invoice_count:
          type: integer
          minimum: 0
        total_amount:
          type: number
        currency:
          type: string

    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string

  responses:
    ValidationError:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Unsupported version or missing resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'