# legacy-api/src/server.ts
import express, { Request, Response } from "express";

type ApiVersion = "v1.1" | "v1.2";

type TemplateRecord = {
  templateToken: string;
  customerId: string;
  expiresAtMs: number;
};

type OrderRecord = {
  orderId: string;
  customerId: string;
  pickupDate: string;
  commodityId: string;
  currency: string;
  lineItems: Array<{
    sku: string;
    quantity: number;
    unitPrice: number;
  }>;
  createdAt: string;
  updatedAt: string;
};

type InvoiceRecord = {
  invoiceId: string;
  customerId: string;
  total: number;
  currency: string;
  issuedAt: string;
  notes?: string;
};

const app = express();
const PORT = Number(process.env.PORT ?? 4011);
const API_KEY = process.env.LEGACY_API_KEY ?? "legacy-test-key";
const TEMPLATE_TTL_MS = 5 * 60 * 1000;

const templates = new Map<string, TemplateRecord>();
const ordersByVersion: Record<ApiVersion, Map<string, OrderRecord>> = {
  "v1.1": new Map<string, OrderRecord>(),
  "v1.2": new Map<string, OrderRecord>()
};
const invoicesByVersion: Record<ApiVersion, Map<string, InvoiceRecord>> = {
  "v1.1": new Map<string, InvoiceRecord>(),
  "v1.2": new Map<string, InvoiceRecord>()
};

seedInvoices();

app.use(express.json());

app.get("/health", (_req: Request, res: Response) => {
  res.status(200).json({ status: "ok" });
});

app.use((req: Request, res: Response, next) => {
  const key = req.header("x-api-key");
  if (key !== API_KEY) {
    return res.status(401).json({
      error: "unauthorized",
      message: "Missing or invalid x-api-key"
    });
  }
  return next();
});

app.get("/api/:version/system/info", (req: Request, res: Response) => {
  const version = parseVersion(req.params.version);
  if (!version) {
    return invalidVersion(res);
  }

  return res.status(200).json({
    system: "Legacy Freight Suite",
    version,
    auth: {
      header: "x-api-key"
    },
    orderCreateFlow: [
      "GET /api/{version}/orders/template?customer_id={id}",
      "PUT /api/{version}/orders/{order_id}"
    ],
    notes:
      version === "v1.1"
        ? "Order payload uses commodity_id"
        : "Order payload uses commodity_code_id"
  });
});

app.get("/api/:version/orders/template", (req: Request, res: Response) => {
  const version = parseVersion(req.params.version);
  if (!version) {
    return invalidVersion(res);
  }

  const customerId = readRequiredString(req.query.customer_id);
  if (!customerId) {
    return res.status(400).json({
      error: "validation_error",
      message: "customer_id is required"
    });
  }

  const templateToken = `tpl_${Math.random().toString(36).slice(2, 12)}`;
  const expiresAtMs = Date.now() + TEMPLATE_TTL_MS;

  templates.set(templateToken, {
    templateToken,
    customerId,
    expiresAtMs
  });

  return res.status(200).json({
    template_token: templateToken,
    customer_id: customerId,
    expires_at: new Date(expiresAtMs).toISOString(),
    default_currency: "USD",
    required_fields:
      version === "v1.1"
        ? ["template_token", "customer_id", "pickup_date", "commodity_id", "line_items"]
        : ["template_token", "customer_id", "pickup_date", "commodity_code_id", "line_items"]
  });
});

app.put("/api/:version/orders/:orderId", (req: Request, res: Response) => {
  const version = parseVersion(req.params.version);
  if (!version) {
    return invalidVersion(res);
  }

  const orderId = readRequiredString(req.params.orderId);
  if (!orderId) {
    return res.status(400).json({
      error: "validation_error",
      message: "orderId path parameter is required"
    });
  }

  const body = req.body as Record<string, unknown>;
  const templateToken = readRequiredString(body.template_token);
  const customerId = readRequiredString(body.customer_id);
  const pickupDate = readRequiredString(body.pickup_date);
  const commodityId =
    version === "v1.1"
      ? readRequiredString(body.commodity_id)
      : readRequiredString(body.commodity_code_id);

  const lineItems = body.line_items;

  if (!templateToken || !customerId || !pickupDate || !commodityId || !isValidLineItems(lineItems)) {
    return res.status(400).json({
      error: "validation_error",
      message:
        version === "v1.1"
          ? "Required: template_token, customer_id, pickup_date, commodity_id, line_items[]"
          : "Required: template_token, customer_id, pickup_date, commodity_code_id, line_items[]"
    });
  }

  const template = templates.get(templateToken);
  if (!template) {
    return res.status(400).json({
      error: "validation_error",
      message: "Invalid template_token"
    });
  }

  if (template.expiresAtMs < Date.now()) {
    return res.status(400).json({
      error: "validation_error",
      message: "template_token has expired"
    });
  }

  if (template.customerId !== customerId) {
    return res.status(400).json({
      error: "validation_error",
      message: "template_token does not match customer_id"
    });
  }

  const store = ordersByVersion[version];
  const now = new Date().toISOString();
  const existing = store.get(orderId);
  const currency = readRequiredString(body.currency) ?? "USD";

  const order: OrderRecord = {
    orderId,
    customerId,
    pickupDate,
    commodityId,
    currency,
    lineItems,
    createdAt: existing?.createdAt ?? now,
    updatedAt: now
  };

  store.set(orderId, order);

  return res.status(existing ? 200 : 201).json(toOrderResponse(order, version));
});

app.get("/api/:version/orders/:orderId", (req: Request, res: Response) => {
  const version = parseVersion(req.params.version);
  if (!version) {
    return invalidVersion(res);
  }

  const orderId = readRequiredString(req.params.orderId);
  if (!orderId) {
    return res.status(400).json({
      error: "validation_error",
      message: "orderId path parameter is required"
    });
  }

  const order = ordersByVersion[version].get(orderId);
  if (!order) {
    return res.status(404).json({
      error: "not_found",
      message: "Order not found"
    });
  }

  return res.status(200).json(toOrderResponse(order, version));
});

app.post("/api/:version/invoices", (req: Request, res: Response) => {
  const version = parseVersion(req.params.version);
  if (!version) {
    return invalidVersion(res);
  }

  const body = req.body as Record<string, unknown>;
  const invoiceId = readRequiredString(body.invoice_id);
  const customerId = readRequiredString(body.customer_id);
  const currency = readRequiredString(body.currency);
  const issuedAt = readRequiredString(body.issued_at);
  const total = readRequiredNumber(body.total);
  const notes = readOptionalString(body.notes);

  if (!invoiceId || !customerId || !currency || !issuedAt || total === undefined) {
    return res.status(400).json({
      error: "validation_error",
      message: "Required: invoice_id, customer_id, total, currency, issued_at"
    });
  }

  const invoice: InvoiceRecord = {
    invoiceId,
    customerId,
    total,
    currency,
    issuedAt,
    notes
  };

  invoicesByVersion[version].set(invoiceId, invoice);

  return res.status(201).json({
    invoice_id: invoice.invoiceId,
    customer_id: invoice.customerId,
    total: invoice.total,
    currency: invoice.currency,
    issued_at: invoice.issuedAt,
    notes: invoice.notes
  });
});

app.get("/api/:version/invoices/:invoiceId", (req: Request, res: Response) => {
  const version = parseVersion(req.params.version);
  if (!version) {
    return invalidVersion(res);
  }

  const invoiceId = readRequiredString(req.params.invoiceId);
  if (!invoiceId) {
    return res.status(400).json({
      error: "validation_error",
      message: "invoiceId path parameter is required"
    });
  }

  const invoice = invoicesByVersion[version].get(invoiceId);
  if (!invoice) {
    return res.status(404).json({
      error: "not_found",
      message: "Invoice not found"
    });
  }

  return res.status(200).json({
    invoice_id: invoice.invoiceId,
    customer_id: invoice.customerId,
    total: invoice.total,
    currency: invoice.currency,
    issued_at: invoice.issuedAt,
    notes: invoice.notes
  });
});

app.get("/api/:version/expenses/summary", (req: Request, res: Response) => {
  const version = parseVersion(req.params.version);
  if (!version) {
    return invalidVersion(res);
  }

  const month = readRequiredString(req.query.month);
  const customerId = readRequiredString(req.query.customer_id);

  if (!month || !/^\d{4}-\d{2}$/.test(month)) {
    return res.status(400).json({
      error: "validation_error",
      message: "month is required and must be in YYYY-MM format"
    });
  }

  if (!customerId) {
    return res.status(400).json({
      error: "validation_error",
      message: "customer_id is required"
    });
  }

  const invoices = [...invoicesByVersion[version].values()]
    .filter((invoice) => invoice.customerId === customerId)
    .filter((invoice) => invoice.issuedAt.startsWith(month));

  const totalAmount = invoices.reduce((sum, invoice) => sum + invoice.total, 0);

  return res.status(200).json({
    customer_id: customerId,
    month,
    invoice_count: invoices.length,
    total_amount: Number(totalAmount.toFixed(2)),
    currency: "USD"
  });
});

app.listen(PORT, () => {
  // eslint-disable-next-line no-console
  console.log(`Legacy fixture API is running on port ${PORT}`);
});

function parseVersion(value: unknown): ApiVersion | null {
  if (value === "v1.1" || value === "v1.2") {
    return value;
  }
  return null;
}

function invalidVersion(res: Response): Response {
  return res.status(404).json({
    error: "not_found",
    message: "Supported versions are v1.1 and v1.2"
  });
}

function isValidLineItems(value: unknown): value is Array<{ sku: string; quantity: number; unitPrice: number }> {
  if (!Array.isArray(value) || value.length === 0) {
    return false;
  }

  return value.every((item) => {
    if (!item || typeof item !== "object") {
      return false;
    }

    const row = item as Record<string, unknown>;
    return (
      typeof row.sku === "string" &&
      row.sku.length > 0 &&
      typeof row.quantity === "number" &&
      row.quantity > 0 &&
      typeof row.unit_price === "number" &&
      row.unit_price >= 0
    );
  });
}

function toOrderResponse(order: OrderRecord, version: ApiVersion): Record<string, unknown> {
  return {
    order_id: order.orderId,
    customer_id: order.customerId,
    pickup_date: order.pickupDate,
    ...(version === "v1.1"
      ? { commodity_id: order.commodityId }
      : { commodity_code_id: order.commodityId }),
    currency: order.currency,
    line_items: order.lineItems.map((item) => ({
      sku: item.sku,
      quantity: item.quantity,
      unit_price: item.unitPrice
    })),
    created_at: order.createdAt,
    updated_at: order.updatedAt
  };
}

function readRequiredString(value: unknown): string | undefined {
  if (typeof value === "string" && value.trim().length > 0) {
    return value.trim();
  }
  return undefined;
}

function readOptionalString(value: unknown): string | undefined {
  if (typeof value === "string") {
    const trimmed = value.trim();
    return trimmed.length > 0 ? trimmed : undefined;
  }
  return undefined;
}

function readRequiredNumber(value: unknown): number | undefined {
  if (typeof value === "number" && Number.isFinite(value)) {
    return value;
  }
  return undefined;
}

function seedInvoices(): void {
  const seedInvoices: InvoiceRecord[] = [
    {
      invoiceId: "inv_1001",
      customerId: "cust_a",
      total: 1200.5,
      currency: "USD",
      issuedAt: "2026-02-05T12:00:00.000Z"
    },
    {
      invoiceId: "inv_1002",
      customerId: "cust_a",
      total: 450.0,
      currency: "USD",
      issuedAt: "2026-02-18T08:30:00.000Z"
    },
    {
      invoiceId: "inv_2001",
      customerId: "cust_b",
      total: 199.99,
      currency: "USD",
      issuedAt: "2026-02-09T14:45:00.000Z"
    }
  ];

  (Object.keys(invoicesByVersion) as ApiVersion[]).forEach((version) => {
    const store = invoicesByVersion[version];
    seedInvoices.forEach((invoice) => {
      store.set(invoice.invoiceId, { ...invoice });
    });
  });
}